requirements.txt
sqlalchemy
psycopg2-binary
pandas
python-dotenv
certifi
requests
beautifulsoup4
selenium
webdriver-manager